final/pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.10",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
